Separate out generate details form.
