+++ apps/admin-portal/src/components/applications/forms/index.ts
@@ -19,3 +19,4 @@
 export * from "./advanced-configurations-form";
 export * from "./inbound-form-factory";
 export * from "./inbound-oidc-form";
+export * from "./general-details-form";
