+++ apps/admin-portal/src/components/applications/general-details-application.tsx
@@ -18,15 +18,14 @@
 
 import { AlertLevels } from "@wso2is/core/models";
 import { addAlert } from "@wso2is/core/store";
-import { Field, Forms, FormValue } from "@wso2is/forms";
-import { ContentLoader, DangerZone, DangerZoneGroup, Hint } from "@wso2is/react-components";
-import React, { FunctionComponent, useState } from "react";
+import { ContentLoader, DangerZone, DangerZoneGroup } from "@wso2is/react-components";
+import React, { FunctionComponent } from "react";
 import { useDispatch } from "react-redux";
-import { Button, Grid, Icon } from "semantic-ui-react";
 import { deleteApplication, updateApplicationDetails } from "../../api";
 import { GlobalConfig } from "../../configs";
 import { history } from "../../helpers";
 import { ApplicationInterface } from "../../models";
+import { GeneralDetailsForm } from "./forms";
 
 /**
  * Proptypes for the applications general details component.
@@ -58,7 +57,6 @@ export const GeneralDetailsApplication: FunctionComponent<GeneralSettingsProps> 
         isLoading
     } = props;
 
-    const [ isDiscoverable, setDiscoverability ] = useState<boolean>(discoverability);
     const dispatch = useDispatch();
 
     const handleApplicationDelete = () => {
@@ -91,19 +89,9 @@ export const GeneralDetailsApplication: FunctionComponent<GeneralSettingsProps> 
             });
     };
 
-    const handleFormSubmit = (values: Map<string, FormValue>) => {
-        const editingApplication: ApplicationInterface = {
-            accessUrl: values.get("accessUrl").toString(),
-            advancedConfigurations: {
-                discoverableByEndUsers: !!values.get("discoverableByEndUsers").includes("discoverable"),
-            },
-            description: values.get("description").toString(),
-            id: appId,
-            imageUrl: values.get("imageUrl").toString(),
-            name: values.get("name").toString(),
-        };
+    const handleFormSubmit = (updatedDetails: ApplicationInterface) => {
 
-        updateApplicationDetails(editingApplication)
+        updateApplicationDetails(updatedDetails)
             .then((response) => {
                 dispatch(addAlert({
                     description: "Successfully updated the application",
@@ -120,102 +108,21 @@ export const GeneralDetailsApplication: FunctionComponent<GeneralSettingsProps> 
             });
     };
 
-    const handleFormValuesOnChange = (isPure: boolean, values: Map<string, FormValue>) => {
-        // Set the discoverability based on the checkbox toggle.
-        if (values.get("discoverableByEndUsers").includes("discoverable") !== isDiscoverable) {
-            setDiscoverability(!!values.get("discoverableByEndUsers").includes("discoverable"));
-        }
-    };
-
     return (
         <>
             {
                 !isLoading
                     ? (
                         <>
-                            <Forms
+                            <GeneralDetailsForm
+                                name={ name }
+                                appId={ appId }
+                                description={ description }
+                                discoverability={ discoverability }
                                 onSubmit={ handleFormSubmit }
-                                onChange={ handleFormValuesOnChange }
-                            >
-                                <Grid>
-                                    <Grid.Row columns={ 1 }>
-                                        <Grid.Column mobile={ 16 } tablet={ 16 } computer={ 8 }>
-                                            <Field
-                                                name="name"
-                                                label="Application Name"
-                                                required={ true }
-                                                requiredErrorMessage="Application name is required"
-                                                placeholder={ name }
-                                                type="text"
-                                                value={ name }
-                                            />
-                                        </Grid.Column>
-                                    </Grid.Row>
-                                    <Grid.Row columns={ 1 }>
-                                        <Grid.Column mobile={ 16 } tablet={ 16 } computer={ 8 }>
-                                            <Field
-                                                name="description"
-                                                label="Description"
-                                                required={ false }
-                                                requiredErrorMessage=""
-                                                placeholder="Enter a description for the application"
-                                                type="textarea"
-                                                value={ description }
-                                            />
-                                        </Grid.Column>
-                                    </Grid.Row>
-                                    <Grid.Row columns={ 1 }>
-                                        <Grid.Column mobile={ 16 } tablet={ 16 } computer={ 8 }>
-                                            <Field
-                                                name="imageUrl"
-                                                label="Application Image"
-                                                required={ false }
-                                                requiredErrorMessage=""
-                                                placeholder="Enter a image url for the application"
-                                                type="text"
-                                                value={ imageUrl }
-                                            />
-                                        </Grid.Column>
-                                    </Grid.Row>
-                                    <Grid.Row columns={ 1 }>
-                                        <Grid.Column mobile={ 16 } tablet={ 16 } computer={ 8 }>
-                                            <Field
-                                                name="discoverableByEndUsers"
-                                                required={ false }
-                                                requiredErrorMessage=""
-                                                type="checkbox"
-                                                children={ [
-                                                    {
-                                                        label: "Discoverable application",
-                                                        value: "discoverable"
-                                                    }
-                                                ] }
-                                                value={ isDiscoverable ? [ "discoverable" ] : [] }
-                                            />
-                                            <Field
-                                                name="accessUrl"
-                                                label="Access URL"
-                                                required={ isDiscoverable }
-                                                requiredErrorMessage={ "A valid access URL needs to be defined for" +
-                                                " an application to be marked as discoverable" }
-                                                placeholder="Enter access url for the application login page"
-                                                type="text"
-                                                value={ accessUrl }
-                                            />
-                                            <Hint>
-                                                Applications flagged as discoverable are visible for end users.
-                                            </Hint>
-                                        </Grid.Column>
-                                    </Grid.Row>
-                                    <Grid.Row columns={ 1 }>
-                                        <Grid.Column mobile={ 16 } tablet={ 16 } computer={ 8 }>
-                                            <Button primary type="submit" size="small" className="form-button">
-                                                Update
-                                            </Button>
-                                        </Grid.Column>
-                                    </Grid.Row>
-                                </Grid>
-                            </Forms>
+                                imageUrl={ imageUrl }
+                                accessUrl={ accessUrl }
+                            />
                             { !(GlobalConfig.doNotDeleteApplications.includes(name)) &&
                             (
                                 <DangerZoneGroup sectionHeader="Danger Zone">
@@ -230,7 +137,7 @@ export const GeneralDetailsApplication: FunctionComponent<GeneralSettingsProps> 
                             }
                         </>
                     )
-                    : <ContentLoader />
+                    : <ContentLoader/>
             }
         </>
     );
